Traditional Sitting Position Versus Hamstring Stretch Position

Initial results from a study to determine whether the traditional sitting position or the hamstring stretch position is superior for administering epidural anesthesia to pregnant women in labor as measured by the number of obstructive (needle to bone) contacts (oc)

EPIDURAL

Format

A data frame with 85 observations on the following 7 variables:

doctor (a factor with levels Dr. A, Dr. B, Dr. C, and Dr. D)

kg (weight in kg of patient)

cm (height in cm of patient)

ease (a factor with levels Difficult, Easy, and Impossible indicating the physicians' assessments of how well bone landmarks could be felt in the patient)

treatment (a factor with levels Hamstring Stretch and Traditional Sitting)

oc (number of obstructive contacts)

complications (a factor with levels Failure - person got dizzy, Failure - too many OCs, None, Paresthesia, and Wet Tap)

Source

Fisher, K. S., Arnholt, A. T., Douglas, M. E., Vandiver, S. L., Nguyen, D. H. 2009. “A Randomized Trial of the Traditional Sitting Position Versus the Hamstring Stretch Position for Labor Epidural Needle Placement.” Journal of Anesthesia & Analgesia, Vol 109, No. 2: 532-534.

References

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. Probability and Statistics with R, Second Edition. Chapman & Hall / CRC.

Examples

xtabs(~ doctor + ease, data = EPIDURAL)

#>        ease
#> doctor  Difficult Easy Impossible
#>   Dr. A         3   19          1
#>   Dr. B        10    7          4
#>   Dr. C         3   18          0
#>   Dr. D         4   13          3

xtabs(~ doctor + factor(ease, levels = c("Easy", "Difficult", "Impossible")), 
data = EPIDURAL)

#>        factor(ease, levels = c("Easy", "Difficult", "Impossible"))
#> doctor  Easy Difficult Impossible
#>   Dr. A   19         3          1
#>   Dr. B    7        10          4
#>   Dr. C   18         3          0
#>   Dr. D   13         4          3