Telephone Call Times

PHONE contains times in minutes of long distance telephone calls during a one month period for a small business.

PHONE

Format

A data frame with 23 observations on the following variable:

call.time (time spent on long distance calls in minutes)

References

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. Probability and Statistics with R, Second Edition. Chapman & Hall / CRC.

Examples

with(data = PHONE, 
SIGN.test(call.time, md = 2.1))
#> 
#> 	One-sample Sign-Test
#> 
#> data:  call.time
#> s = 11, p-value = 1
#> alternative hypothesis: true median is not equal to 2.1
#> 95 percent confidence interval:
#>  0.7261939 5.4952244
#> sample estimates:
#> median of x 
#>         2.1 
#> 
#> Achieved and Interpolated Confidence Intervals: 
#> 
#>                   Conf.Level L.E.pt U.E.pt
#> Lower Achieved CI     0.9069 0.8000 5.2000
#> Interpolated CI       0.9500 0.7262 5.4952
#> Upper Achieved CI     0.9653 0.7000 5.6000
#>