Workstation Comparison

To compare the speed differences between two different brands of workstations (Sun and Digital), the times each brand took to complete complex simulations were recorded. Five complex simulations were selected, and the five selected simulations were run on both workstations. The resulting times in minutes for the five simulations are stored in data frame SUNDIG.

SUNDIG

Format

A data frame with 5 observations on the following 3 variables:

sun (time in seconds for a Sun workstation to complete a simulation)

digital (time in seconds for a Digital workstation to complete a simulation)

difference (difference between sun and digital)

References

Ugarte, M. D., Militino, A. F., and Arnholt, A. T. 2015. Probability and Statistics with R, Second Edition. Chapman & Hall / CRC.

Examples

with(data = SUNDIG, t.test(sun, digital, paired=TRUE)$conf)

#> [1] 0.6100548 9.3899452
#> attr(,"conf.level")
#> [1] 0.95