Simple Random Sample

Computes all possible samples from a given population using simple random sampling

srs(popvalues, n)

Arguments

popvalues

are values of the population. NAs and Infs are allowed but will be removed from the population.

n

the sample size

Value

The function srs() returns a matrix containing the possible simple random samples of size n taken from a population of finite values popvalues.

Details

If non-finite values are entered as part of the population, they are removed; and the returned simple random sample computed is based on the remaining finite values.

See also

combn

Examples

srs(popvalues = c(5, 8, 3, NA, Inf), n = 2)

#>      [,1] [,2]
#> [1,]    5    8
#> [2,]    5    3
#> [3,]    8    3
#> 

Author
Alan T. Arnholt <arnholtat@appstate.edu>